feat: combine changelog and about dialogs
