+++ web/src/components/modals/AboutModal/ChangeLog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ChangeLog';
