+++ web/src/components/modals/AboutModal/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './AboutModal';
