--- web/src/components/modals/ChangeLogModal.css
@@ -1,3 +1,0 @@
-.ChangeLogModal__issue-url:after {
-  content: " - ";
-}
